Control 패널에 대한 로그인 체크로직 변경
- 자바스크립트단에서 API를 호출하는 것이 아닌, PHP 상에서 모듈을 불러와 로그인 체크하여 메시지 반환하는 식으로

+++ process/clearAll.php
@@ -13,6 +13,15 @@
 
 if (defined('__IM__') == false) exit;
 
+$memberModule = $this->IM->getModule('member');
+$logged = $memberModule->isLogged();
+
+if (!$logged) {
+    $results->success = false;
+    $results->message = 'NOT LOGGED';
+    return;
+}
+
 $success = $this->db()->delete($this->table->todolist)->execute();
 
 $results->success = true;
